API: Implemented bet spread; round bets to min increment

+++ Backend/src/main/kotlin/com/cardcounting/backend/dto/StrategyResponse.kt
@@ -1,7 +1,7 @@
 package com.cardcounting.backend.dto
 
 data class StrategyResponse(
-    val betSize: Int,
+    val betSize: Double,
     val hardTable: Map<String, String>,
     val softTable: Map<String, String>,
     val splitTable: Map<String, String>
